전남대 BE_모아림 6주차 과제 (3단계) (#387)
* Update README.md

* refactor: 특정 도메인만 허용하도록 WebConfig 수정

* refactor: Product 의 price를 int로 수정

* feat: Member 도메인에 point 추가

* feat: 포인터 차감, 포인트 생성 기능 추가

* feat: Order에 포인트 추가

* feat: OrderService 포인트 기능 구현

* refactor: earnPoint() >> addPoint() 함수명 변경

* fix: 함수명 변경 반영

+++ src/main/java/gift/domain/Product.java
@@ -23,7 +23,7 @@ public class Product {
     private String name;
 
     @Column(nullable = false)
-    private double price;
+    private int price;
 
     @Column(name = "image_url", nullable = false)
     private String imageUrl;
@@ -35,14 +35,14 @@ public class Product {
     protected Product() {
     }
 
-    public Product(String name, double price, String imageUrl, Category category) {
+    public Product(String name, int price, String imageUrl, Category category) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
         this.category = category;
     }
 
-    public Product(Long id, String name, double price, String imageUrl, Category category) {
+    public Product(Long id, String name, int price, String imageUrl, Category category) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -62,7 +62,7 @@ public class Product {
         return name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
@@ -76,7 +76,7 @@ public class Product {
         return categoryDto;
     }
 
-    public void update(String name, double price, String imageUrl, Category category) {
+    public void update(String name, int price, String imageUrl, Category category) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
