+++ src/main/java/gift/dto/ProductDto.java
@@ -4,18 +4,18 @@ public class ProductDto {
 
     private final Long id;
     private final String name;
-    private final double price;
+    private final int price;
     private final String imageUrl;
     private String CategoryName;
 
-    public ProductDto(Long id, String name, double price, String imageUrl) {
+    public ProductDto(Long id, String name, int price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
     }
 
-    public ProductDto(Long id, String name, double price, String imageUrl, String categoryName) {
+    public ProductDto(Long id, String name, int price, String imageUrl, String categoryName) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -31,7 +31,7 @@ public class ProductDto {
         return name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
